requirements.txt
nicegui
sqlmodel
sqlalchemy
aiosqlite
pydantic>=2
httpx
apscheduler
python-dotenv
requests
